requirements.txt
numpy==1.17.4
Pillow==8.4.0
transforms3d==0.3.1
